Start working the database scripts
