--- data/user_service.js
@@ -1,8 +1,0 @@
-const USERSERVICE = [
-  {
-    user_id: 1,
-    service_id: 2,
-  },
-];
-
-export default USERSERVICE;
